adk_backend/requirements.txt
# Core Python + API
fastapi>=0.110,<0.120
uvicorn[standard]>=0.29,<0.30

# ADK / CrewAI Core
crewai==0.130.0
langchain>=0.2.0,<0.3.0
openai>=1.68.2,<2.0.0



# YAML parsing
PyYAML==6.0.1

# Pydantic for request validation
pydantic==2.7.1

# Typing
typing_extensions>=4.5.0,<5.0.0

# Google Cloud / ADK integration
google-cloud-aiplatform>=1.97.0,<2.0.0
google-cloud-storage>=2.16.0,<3.0.0
google-cloud-logging>=3.10.0,<4.0.0
google-cloud-secret-manager>=2.18.3,<3.0.0
google-cloud-pubsub>=2.21.0,<3.0.0

# Data support
pandas>=2.2.2,<3.0.0
python-dotenv>=1.0.1,<2.0.0

# Compatibility
protobuf>=4.25.0,<5.0.0
